requirements.txt
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
opencv-python~=4.6.0.66
numpy~=1.21.6
pandas~=1.3.5
matplotlib~=3.5.3
scipy==1.7.3
wget~=3.2
Pillow~=9.3.0
requests~=2.28.1